Extract AchievementTierBadge into own component
Move AchievementTierBadge out of achievement-card into a new shared achievement-tier-badge component with its own index and styles. Scope and dedupe styling by adding .tier--<tier> modifiers and removing duplicate rules from achievement-card/styles.module.sass. Update imports/usages: re-export the badge and update admin achievements page to use the new component (simplify formatter lambdas and remove an extra <strong> wrapper around the title). This refactor improves reuse and isolates tier-specific styling.
