+++ client/components/shared/achievement-tier-badge/index.ts
@@ -1,0 +1,1 @@
+export { AchievementTierBadge } from './AchievementTierBadge'
